requirements.txt
gradio
Pillow
torch
transformers
numpy
matplotlib